+++ src/openmodel/src/primitives/xform.rs
@@ -11,7 +11,7 @@ use serde_json::Value;
 #[derive(Debug, Clone, Deserialize)]
 pub struct Xform {
     /// The matrix elements stored in column-major order as a flattened array
-    pub m: [f64; 16],
+    pub m: [f32; 16],
 }
 
 impl Xform {
@@ -20,7 +20,7 @@ impl Xform {
     /// # Arguments
     ///
     /// * `value` - The value to initialize all elements with
-    pub fn new(value: f64) -> Self {
+    pub fn new(value: f32) -> Self {
         Xform { m: [value; 16] }
     }
 
@@ -64,7 +64,7 @@ impl Xform {
     /// assert_eq!(transformed.y, 4.0);
     /// assert_eq!(transformed.z, 5.0);
     /// ```
-    pub fn translation(tx: f64, ty: f64, tz: f64) -> Self {
+    pub fn translation(tx: f32, ty: f32, tz: f32) -> Self {
         let mut xform = Self::identity();
         xform.m[12] = tx;
         xform.m[13] = ty;
@@ -90,7 +90,7 @@ impl Xform {
     /// assert_eq!(transformed.y, 3.0);
     /// assert_eq!(transformed.z, 4.0);
     /// ```
-    pub fn scaling(sx: f64, sy: f64, sz: f64) -> Self {
+    pub fn scaling(sx: f32, sy: f32, sz: f32) -> Self {
         let mut xform = Self::identity();
         xform.m[0] = sx;
         xform.m[5] = sy;
@@ -103,7 +103,7 @@ impl Xform {
     /// # Arguments
     ///
     /// * `angle_radians` - Rotation angle in radians
-    pub fn rotation_x(angle_radians: f64) -> Self {
+    pub fn rotation_x(angle_radians: f32) -> Self {
         let mut xform = Self::identity();
         let cos_angle = angle_radians.cos();
         let sin_angle = angle_radians.sin();
@@ -121,7 +121,7 @@ impl Xform {
     /// # Arguments
     ///
     /// * `angle_radians` - Rotation angle in radians
-    pub fn rotation_y(angle_radians: f64) -> Self {
+    pub fn rotation_y(angle_radians: f32) -> Self {
         let mut xform = Self::identity();
         let cos_angle = angle_radians.cos();
         let sin_angle = angle_radians.sin();
@@ -139,7 +139,7 @@ impl Xform {
     /// # Arguments
     ///
     /// * `angle_radians` - Rotation angle in radians
-    pub fn rotation_z(angle_radians: f64) -> Self {
+    pub fn rotation_z(angle_radians: f32) -> Self {
         let mut xform = Self::identity();
         let cos_angle = angle_radians.cos();
         let sin_angle = angle_radians.sin();
@@ -158,7 +158,7 @@ impl Xform {
     ///
     /// * `axis` - The axis of rotation (should be normalized)
     /// * `angle_radians` - Rotation angle in radians
-    pub fn rotation(axis: &Vector, angle_radians: f64) -> Self {
+    pub fn rotation(axis: &Vector, angle_radians: f32) -> Self {
         assert!((axis.length() - 1.0).abs() < 1e-6, "Axis must be normalized");
         
         let mut xform = Self::identity();
@@ -398,7 +398,7 @@ impl Default for Xform {
 
 // Implement Index trait for accessing matrix elements with [(row, col)] syntax
 impl Index<(usize, usize)> for Xform {
-    type Output = f64;
+    type Output = f32;
 
     fn index(&self, idx: (usize, usize)) -> &Self::Output {
         let (row, col) = idx;
@@ -459,7 +459,7 @@ impl MulAssign for Xform {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::f64::consts::PI;
+    use std::f32::consts::PI;
 
     #[test]
     fn test_identity() {
@@ -621,7 +621,7 @@ impl FromJsonData for Xform {
                 let mut m = [0.0; 16];
                 for (i, val) in m_array.iter().enumerate() {
                     if let Some(f) = val.as_f64() {
-                        m[i] = f;
+                        m[i] = f as f32;
                     } else {
                         return None;
                     }
